created seperate file for plategen functions
